server/requirements.txt
torch
torchaudio
openai-whisper @ git+https://github.com/openai/whisper.git
flask
pymongo
python-dotenv